requirements.txt
moviepy
openai-whisper
python-dotenv